refactor(markdown): replacer
